Fix same answer loaded for all ubcpi in a unit.
If multiple ubcpi xblocks are in a unit, they all load the exact same
answer. This is due to the selector for the answer loading not being
restricted to just the single xblock it belongs to.

Also added default values for question and options. They were causing an
exception in newer versions of edx.