Present other students' answers for revise stage.
Other students' answers are now shown in the revise stage. Only the
option that the other student selected is shown.

Student's answers are put into a limited size pool. The pool is stored
as an xblock field. This is meant to provide faster access than having
to read every submission from students.

The current pool management scheme is simple. All submissions are
entered into the pool until the pool is filled up. Once the pool is
filled up, a random entry selected for eviction to make room for new
submissions.

The answer selection for presenting to the user is also simple. For
every option, we select a random student who selected that option.