Add random seed to generate same set of answers for students
When the student finished the first answer, a set of random selected
answers from other students are presented to them. In order to generate
the same set of answers when they refresh the page, the student id is
used as a seed for generating the random numbers. This approach may have
a flaw that when the answer pool size changed, they may get a different
set of answers. However, assuming student doesn't wait long enough to
allow pool size changes, we use this for current version.