ADD ability to use images as options
- instructors can now use images instead of text as options for their
problem
	- the images have to be uploaded via Edx File & Upload
	- the "studio url" is inserted into the text box in place
	of text
	- the option type will also have to be changed to "image"
- I've also modified the html to display the images instead of the
images' url on the "show others' answers" and stats screens
- The studio url is converted to the "assets" url on the fly so that
instructors will not get confused if they go back and see a different
url in the edit interface

- A minor css overhaul is needed so that some text can be inline
with the images better.
- Also the option text are not appearing below the bar graphs anymore
if any images are used
- Most importantly, I have modified option from a string to an object
with the name and type attributes. This will break any problems created
before this commit.