Solve the conflicts when two or more xblocks on the same page
in order to support multiple same apps on the same page but
under different elements, e.g. multiple xblocks on the same
page, the data and URLs have to be passed into scope. We can
not use angular constant or value because they are global and
can be override by the second app. They can be passed with element,
which is converted to a angular value as $rootElement and can
be injected as dependencies, through angular.bootstrap. Element
is unique for each app.