Improve scenario with no correct answer
In studio view, change "No correct answer" to "n/a". Insert red
asterisks to indicate required fields and add "optional" for
Explanation field on-select of "n/a" for Correct Answer. Make
Explanation/rationale field optional if instructor chooses to not
specify a correct answer.

In student view, add "RATIONALE" heading for final section if no
correct answer is specified and a rationale is entered by the
instructor. Display "CORRECT ANSWER" if a correct answer is specified.

Modify tests to accommodate new changes.