Fix bug with incorrect colouring of final answer
If the instructor has chosen to not specify a correct answer, the
student's revised answer will be black (and not red).

Change "RATIONALE" heading to "INSTRUCTOR'S EXPLANATION" in final step,
for the case where the instructor has chosen to not specify a correct
answer and has provided a value in the explanation/rationale field.