Improve scenario with no correct answer (#106)
* Allow setup of a question without a correct answer

An instructor can now specify that a PI question has no correct answer.
The statistics graphs do not indicate a correct answer if none was
specified.

The student's response is also no longer indicated as incorrect (ie
rendered in red) for this case.

Added tests and updated dependencies based on what edX currently uses.

* Improve scenario with no correct answer

In studio view, change "No correct answer" to "n/a". Insert red
asterisks to indicate required fields and add "optional" for
Explanation field on-select of "n/a" for Correct Answer. Make
Explanation/rationale field optional if instructor chooses to not
specify a correct answer.

In student view, add "RATIONALE" heading for final section if no
correct answer is specified and a rationale is entered by the
instructor. Display "CORRECT ANSWER" if a correct answer is specified.

Fix bug where progress bar showing in studio view.

Modify tests to accommodate new changes.

* Fix bug with incorrect colouring of final answer

If the instructor has chosen to not specify a correct answer, the
student's revised answer will be black (and not red).

Change "RATIONALE" heading to "INSTRUCTOR'S EXPLANATION" in final step,
for the case where the instructor has chosen to not specify a correct
answer and has provided a value in the explanation/rationale field.