Use locale name instead of language name (#163)
There is a problem with the translations (see #159)

The translations catalog(ubcpi/static/js/src/translations.js) is generated using locale names and translations from transifex.

The django.utils.translation.get_language() function returns the language code (i.e. es-419), but since in the catalog we are using locale names(i.e. es_419), the strings are not being translated as expected.

With this change, now we are using the locale names and it works fine. I tested this with spanish